backend/requirements.txt

Flask==3.1.2
flask_cors==5.0.1
Werkzeug==3.1.3


neo4j==5.28.2


torch==2.8.0
torchvision>=0.10.0
transformers==4.55.2


spacy==3.8.5


Pillow==11.3.0
opencv-python>=4.5.0
easyocr>=1.6.0


pandas==2.3.2
numpy>=1.21.0
openpyxl>=3.0.9


requests>=2.25.0
urllib3>=1.26.0


python-pptx>=0.6.21
PyMuPDF>=1.20.0
pdfplumber>=0.7.0
camelot-py[cv]>=0.10.1
tabula-py>=2.3.0
pdfminer.six>=20211012


chardet==5.2.0
lxml>=4.6.0
beautifulsoup4>=4.9.0


scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0


tqdm>=4.62.0
python-dotenv>=0.19.0
click>=8.0.0
colorama>=0.4.4
loguru>=0.6.0


jsonschema>=4.0.0


pyyaml>=6.0


concurrent-futures>=3.1.1


psutil>=5.8.0


pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
jupyter>=1.0.0
notebook>=6.4.0
